moved coastline and borders to geography.py
